% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dryclean.R
\name{apg_progect}
\alias{apg_progect}
\alias{apg_project}
\title{Accelerated Proximal Gradient based projection method}
\usage{
apg_project(m.vec, U, lambda1, lambda2)
}
\arguments{
\item{m.vec}{vector of GC corrected coverage data of sample in question}

\item{U}{the basis of low rank subspace}

\item{lambda1}{tuning parameter}

\item{lambda2}{tuning parameter}
}
\value{
list with s and v vectors
}
\description{
project new sample into the burnin space
solving projection by Accelerated Proximal Gradient
min_{v, s} 0.5*|m-Uv-s|_2^2 + 0.5*lambda1*|v|^2 + lambda2*|s|_1
}
\author{
Aditya Deshpande
}
\keyword{internal}
