% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dryclean.R
\name{balance_X}
\alias{balance_X}
\title{Balances X chr and takes PAR region under considersation}
\usage{
balance_X(this.cov, build = build)
}
\arguments{
\item{this.cov}{coverage input}

\item{build}{genome build to define PAR region in chromosome X}
}
\value{
median normalized X chromosome that takes care of single copy
}
\description{
Balances X chr and takes PAR region under considersation
}
\author{
Aditya Deshpande
}
\keyword{internal}
