% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dryclean.R
\name{dryclean}
\alias{dryclean}
\title{dryclean}
\description{
dryclean R6 class storing all methods and values necessary for "drycleaning"
}
\details{
Add more details
}
\author{
Aditya Deshpande <asd3002@med.cornell.edu>, Sebastian Brylka <sebastian.brylka@nyulangone.org>
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-dryclean-new}{\code{dryclean$new()}}
\item \href{#method-dryclean-clean}{\code{dryclean$clean()}}
\item \href{#method-dryclean-get_history}{\code{dryclean$get_history()}}
\item \href{#method-dryclean-get_mismatch}{\code{dryclean$get_mismatch()}}
\item \href{#method-dryclean-clone}{\code{dryclean$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-dryclean-new"></a>}}
\if{latex}{\out{\hypertarget{method-dryclean-new}{}}}
\subsection{Method \code{new()}}{
Initialize dryclean object. Authors: Aditya Deshpande, Sebastian Brylka
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{dryclean$new(pon)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{pon}}{PON object}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-dryclean-clean"></a>}}
\if{latex}{\out{\hypertarget{method-dryclean-clean}{}}}
\subsection{Method \code{clean()}}{
Function begins the online rPCA process. Use this function if you performed batch rPCA on samples as whole without dividing into chromosomes. For exomes and whole genomes where number of normal samples are small (<=100). Author: Aditya Deshpande
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{dryclean$clean(
  cov,
  center = TRUE,
  cbs = FALSE,
  cnsignif = 1e-05,
  mc.cores = 1,
  verbose = TRUE,
  use.blacklist = FALSE,
  blacklist_path = NA,
  germline.filter = FALSE,
  field = "reads.corrected",
  testing = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{cov}}{character path to the granges coverage file to be drycleaned}

\item{\code{center}}{boolean (default == TRUE) whether to center the coverage before drycleaning}

\item{\code{cbs}}{boolean (default == FALSE) whether to perform cbs on the drycleaned coverage}

\item{\code{cnsignif}}{integer (default = 1e-5) the significance levels for the tests in cbs to accept change-points}

\item{\code{mc.cores}}{interger (default == 1) number of cores to use for parallelization}

\item{\code{verbose}}{boolean (default == TRUE) outputs progress}

\item{\code{use.blacklist}}{boolean (default = FALSE) whether to exclude off-target markers in case of Exomes or targeted sequencing. If set to TRUE, needs a GRange marking if each marker is set to be excluded or not or will use a default mask}

\item{\code{blacklist_path}}{character (default = NA) if use.blacklist == TRUE, path a GRanges object marking if each marker is set to be excluded or not}

\item{\code{germline.filter}}{boolean (default == FALSE) if germline markers need to be removed from decomposition}

\item{\code{field}}{character (default == "reads.corrected") field to use for processing}

\item{\code{testing}}{boolean (default = FALSE) DO NOT CHANGE}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
The function begins the online rPCA process. It is the wrapper that takes in GRanges and outputs GRanges with decomposition
}

\subsection{Returns}{
Drycleaned coverage in GRanges format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-dryclean-get_history"></a>}}
\if{latex}{\out{\hypertarget{method-dryclean-get_history}{}}}
\subsection{Method \code{get_history()}}{
Function returns the history of the dryclean object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{dryclean$get_history()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
Prints the history of the dryclean object as data table
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-dryclean-get_mismatch"></a>}}
\if{latex}{\out{\hypertarget{method-dryclean-get_mismatch}{}}}
\subsection{Method \code{get_mismatch()}}{
Function returns the data table with mismatching in seqlengths between pon and coverage
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{dryclean$get_mismatch()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
Returns the data table with mismatching chromosomes of pon and coverage
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-dryclean-clone"></a>}}
\if{latex}{\out{\hypertarget{method-dryclean-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{dryclean$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
