% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dryclean.R
\name{identify_germline}
\alias{identify_germline}
\title{This is first phase in preparing the Panel of Normals (PON)
that will be used for online decomposition}
\usage{
identify_germline(
  normal.table.path = NA,
  signal.thresh = 0.5,
  pct.thresh = 0.98,
  verbose = TRUE,
  save.grm = FALSE,
  path.to.save = NA,
  num.cores = 1
)
}
\arguments{
\item{normal.table.path}{character path to data.table containing two columns "sample", "normal_cov" and additional column "decomposed_cov" that contains drycleaned normal sample outputs. See manual for details.}

\item{signal.thresh}{numeric (default == 0.5). This is the threshold to be used to identify an amplification (markers with signal intensity > 0.5) or deletions (markers with signal intensity < -0.5) in log space from dryclean outputs.}

\item{pct.thresh}{numeric (default == 0.98). Proportion of samples in which a given marker is free of germline event.}

\item{verbose}{boolean (default == TRUE). Outputs progress.}

\item{save.grm}{boolean (default == FALSE). If the germline list needs to be saved.}

\item{path.to.save}{charater (default == NA). Path to save the germline list created if save.grm == TRUE.}

\item{num.cores}{interger (default == 1). Number of cores to use for parallelization.}
}
\value{
A GRange object with a metadata field annotating germline markers.
}
\description{
This function takes in gRanges outputs from fragCounter and extracts GC corrected
read count data and carries rPCA decomposition on the matrix thus created
}
\author{
Aditya Deshpande
}
