% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dryclean.R
\name{prep_cov}
\alias{prep_cov}
\title{function to prepare covearge data for decomposition}
\usage{
prep_cov(m.vec = m.vec)
}
\arguments{
\item{m.vec}{GRanges object containing GC corrected reads as a cloumn in metadata}
}
\value{
matrix with processed coverage data
}
\description{
preapares the GC corrected coverage data for decomposition
}
\author{
Aditya Deshpande
}
\keyword{internal}
