% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dryclean.R
\name{prepare_solvent_phase1}
\alias{prepare_solvent_phase1}
\title{This is first phase in preparing the Panel of Normals
that will be used for online decomposition}
\usage{
prepare_solvent_phase1(normal_table_path = NA, mc.cores = 1)
}
\arguments{
\item{mc.cores}{interger (default == 1). Number of cores to use for parallelization}

\item{normal_table}{character path to data.table containing two columns "pair" and "normal_cov". See manual for details}
}
\value{
list with two decompose matrices: L and S.
}
\description{
This function takes in gRanges outputs from fragCounter and extracts GC corrected
read count data and carrie rPCA decomposition on the matrix thus created
}
\author{
Aditya Deshpande
}
