% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dryclean.R
\name{prepare_solvent_phase2}
\alias{prepare_solvent_phase2}
\title{This is second phase in preparing the Panel of Normals
that will be used for online decomposition}
\usage{
prepare_solvent_phase2(solvent, k = NA, mc.cores = cores)
}
\arguments{
\item{solvent}{list with two decompose matrices: L and S.}

\item{k}{integer (default == NA). This is the approximate rank of PON matrix.}

\item{mc.cores}{interger (default == 1). Number of cores to use for parallelization}
}
\value{
list with two decompose matrices: L and S in addition to left, right sigulare vectors and k singular values.
}
\description{
This function takes in list of rpCA decomposed PON and 
carries randomized SVD on subspace matrix L
}
\author{
Aditya Deshpande
}
