% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dryclean.R
\name{thresh}
\alias{thresh}
\title{Internal shrinkage function for values in S vector}
\usage{
thresh(x, mu)
}
\arguments{
\item{x}{vector to which shrinkage is to applied}

\item{mu}{shrinkage parameter}
}
\value{
regularized vector y
}
\description{
Internal shrinkage function for values in S vector:
y = sgn(x)max(|x| - mu, 0)
}
\author{
Aditya Deshpande
}
\keyword{internal}
