% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dryclean.R
\name{update_cols}
\alias{update_cols}
\title{subspace update based on projection of sample in question}
\usage{
update_cols(U, A, B, lambda1)
}
\arguments{
\item{U}{the basis of low rank subspace to be updated}

\item{A, }{B matrices to uodate subspace basis}

\item{lambda1}{tuning parameter}
}
\value{
updated basis U of subspace
}
\description{
Internal function that updates basis of low rank subspace after
projecting sample in question by warm restart
}
\author{
Aditya Deshpande
}
\keyword{internal}
