% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{validate_pon_paths}
\alias{validate_pon_paths}
\title{validate_pon_paths}
\usage{
validate_pon_paths(
  pon_paths,
  field,
  target_resolution,
  cores,
  all.chr = c(1:22, "X", "Y"),
  verbose = TRUE
)
}
\arguments{
\item{pon_paths}{A character vector of file paths to PON files.}
}
\value{
A character vector of validated PON file paths.
}
\description{
validate_pon_paths
}
\details{
Checks that the provided Panel of Normals (PON) file paths are valid and accessible. This function validates that each file path supplied in the input is a well-formed path and that the file exists. It raises an informative error if any path does not meet the required criteria for subsequent processing. Returns as list of rebinned coverages. The function performs necessary checks for file existence and proper format. Ensure that the provided paths correspond to files that are accessible in the current environment.
}
\examples{
\dontrun{
  # Validate PON file paths
  valid_paths <- validate_pon_paths(c("/path/to/pon1.txt", "/path/to/pon2.txt"))
}
}
