% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dryclean.R
\name{wash_cycle}
\alias{wash_cycle}
\title{function to perform online rPCA descomposition}
\usage{
wash_cycle(m.vec, L.burnin, S.burnin, r, N, U.hat, V.hat, sigma.hat,
  lambda1 = NA, lambda2 = NA, decomp = FALSE, verbose = TRUE)
}
\arguments{
\item{m.vec}{numeric vector of length m. Vector of GC corrected coverage data of sample in question}

\item{L.burnin}{(m  markers x n samples) numeric matrix. L matrix of panel of normals after batch rPCA decomposition}

\item{S.burnin}{(m  markers x n samples) numeric matrix. S matrix of panel of normals after batch rPCA decomposition}

\item{r}{integer. Estimated rank of panel of normals after batch rPCA decomposition}

\item{N}{(m  markers x n samples) numeric matrix. Matrix of GC corrected coverage data of panel of normals for batch rPCA decomposition}

\item{U.hat}{(m  markers x n samples) numeric matrix. Right singular matrix of L.burnin}

\item{V.hat}{(m  markers x n samples) numeric matrix. Left singular matrix of L.burnin}

\item{sigma.hat}{numeric vector of length r. Singular values of L.burnin}

\item{lambda1, }{lambda2 integer. Tuning parameter (default = 1/(sqrt(no. of markers in input vector))}

\item{decomp}{boolean (default = FALSE). If TRUE, carry batch rPCA decomposition on N matrix}
}
\value{
S and L numeric vectors each of length m. Vectors for sample in question
}
\description{
Function that calls the online rPCA on sample under question and updates subspace
basis and does decomposition
}
\author{
Aditya Deshpande
}
