% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{apg_progect}
\alias{apg_progect}
\alias{apg_project}
\title{Accelerated Proximal Gradient based projection method}
\usage{
apg_project(m.vec, U, lambda1, lambda2)
}
\arguments{
\item{m.vec}{numeric vector of length m. Vector of GC corrected coverage data of sample in question}

\item{U}{(m  markers x n samples) numeric matrix. The basis of low rank subspace. The dimensions are same as burnin matrix}

\item{lambda1}{integer. Tuning parameter (default = 1/(sqrt(no. of markers in input vector))}

\item{lambda2}{integer. Tuning parameter (default = 1/(sqrt(no. of markers in input vector))}
}
\value{
list with s and v vectors
}
\description{
project new sample into the burnin space
solving projection by Accelerated Proximal Gradient
}
\author{
Aditya Deshpande
}
\keyword{internal}
