% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{prep_cov}
\alias{prep_cov}
\title{function to prepare covearge data for decomposition}
\usage{
prep_cov(m.vec = m.vec, use.blacklist = FALSE, blacklist = NA)
}
\arguments{
\item{m.vec}{GRanges object. GRanges containing GC corrected reads as a cloumn in metadata}

\item{blacklist, }{boolean (default == FALSE). Whether to exclude off-target markers in case of Exomes or targeted sequqnecing. If set to TRUE, needs a GRange marking if each marker is set to be excluded or not.}

\item{burnin.samples.path, }{character (default = NA). Path to balcklist markers file}
}
\value{
vector of length m with processed coverage data
}
\description{
prepares the GC corrected coverage data for decomposition
}
\author{
Aditya Deshpande
}
\keyword{internal}
