% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dryclean.R
\name{start_wash_cycle}
\alias{start_wash_cycle}
\title{function begins the online rPCA process and parallelizes the chromosmes for speed.}
\usage{
start_wash_cycle(cov, mc.cores = 1, burnin.samples.path = NA,
  verbose = TRUE, chr = NA, whole_genome = TRUE)
}
\arguments{
\item{cov}{GRanges object containig the GC corrected cov data outputed from fragCounter. Needs metadata with header "reads.corrected"}

\item{mc.cores}{interger. Number of cores to use for parallelization}

\item{burnin.samples.path}{string. Path to burnin samples for each chromosome}

\item{chr}{string. if a single chromosome is to be processed, name of chromosome. Requires whole_genome = FALSE}

\item{whole_genome}{boolean (default = TRUE). For processing chromosome or whole genome}
}
\value{
GRange object with decomposition
}
\description{
function begins the online rPCA process and parallelizes the chromosmes for speed.  
It is the wrapper that takes in GRanges and outputs GRanges with decomposition
}
\author{
Aditya Deshpande
}
