% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dryclean.R
\name{start_wash_cycle}
\alias{start_wash_cycle}
\title{function begins the online rPCA process. Use this function if you performed batch rPCA on samples as whole withiout dividing into chromosomes.For exomes and whole genomes where number of normal samples are small (<=100).}
\usage{
start_wash_cycle(
  cov,
  mc.cores = 1,
  detergent.pon.path = NA,
  verbose = TRUE,
  whole_genome = TRUE,
  use.blacklist = FALSE,
  chr = NA,
  germline.filter = FALSE,
  germline.file = NA,
  field = "reads.corrected",
  is.human = TRUE
)
}
\arguments{
\item{cov}{GRanges object containig the GC corrected cov data outputed from fragCounter. Needs metadata with header "reads.corrected".}

\item{mc.cores}{interger (default == 1). Number of cores to use for parallelization.}

\item{detergent.pon.path}{string. Path to pon/detergent genrated using normal samples.}

\item{verbose}{boolean (default == TRUE). Outputs progress.}

\item{whole_genome}{boolean (default = TRUE). For this function always set this parameter to TRUE.}

\item{use.blacklist}{boolean (default = FALSE). Whether to exclude off-target markers in case of Exomes or targeted sequqnecing. If set to TRUE, needs a GRange marking if each marker is set to be excluded or not.}

\item{chr}{integer (default == NA). Depricated. Can be used to decompose a single chromosome.}

\item{germline.filter}{boolean (default == TRUE). If germline markers need to be removed from decomposition.}

\item{germline.file}{character (default == NA). Path to file with germline markers.}

\item{field}{character (default == "reads.corrected"). Field to use for processing.}

\item{is.human}{boolean (default == TRUE). Organism type.}
}
\value{
GRange object with decomposition
}
\description{
function begins the online rPCA process.  
It is the wrapper that takes in GRanges and outputs GRanges with decomposition
}
\author{
Aditya Deshpande
}
