% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dryclean.R
\name{thresh}
\alias{thresh}
\title{Internal shrinkage function for values in S vector}
\usage{
thresh(x, mu)
}
\arguments{
\item{x}{numeric vector of length == number of markers in the sample. Vector to which shrinkage is to be applied.}

\item{mu}{integer. Shrinkage parameter to use (default = lambda2 = 1/(sqrt(no. of markers in input vector)))}
}
\value{
numeric vector. Regularized vector y
}
\description{
Internal shrinkage function for values in S vector:
y = sgn(x)max(|x| - mu, 0)
}
\author{
Aditya Deshpande
}
\keyword{internal}
