% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dryclean.R
\name{update_cols}
\alias{update_cols}
\title{subspace update based on projection of sample in question}
\usage{
update_cols(U, A, B, lambda1)
}
\arguments{
\item{U}{(m  markers x n samples) numeric matrix. The basis of low rank subspace. The dimensions are same as burnin matrix}

\item{A, }{B m x n numeric matrices. Matrices to update subspace basis}

\item{lambda1}{integer. Tuning parameter (default = 1/(sqrt(no. of markers in input vector))}
}
\value{
U (m  markers x n samples) numeric matrix. updated basis U of subspace
}
\description{
Internal function that updates basis of low rank subspace after
projecting sample in question by warm restart
}
\author{
Aditya Deshpande
}
\keyword{internal}
