% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dryclean.R
\name{wash_cycle}
\alias{wash_cycle}
\title{function to perform online rPCA descomposition}
\usage{
wash_cycle(m.vec, L.burnin, S.burnin, r, N, U.hat, V.hat, sigma.hat,
  lambda1 = NA, lambda2 = NA, decomp = FALSE, verbose = TRUE)
}
\arguments{
\item{m.vec}{vector of GC corrected coverage values of sample under question}

\item{L.burnin}{L matrix of panel of normals after batch rPCA decomposition}

\item{S.burnin}{S matrix of panel of normals after batch rPCA decomposition}

\item{r}{estimated rank of panel of normals after batch rPCA decomposition}

\item{N}{matrix of GC corrected coverage data of panel of normals for batch rPCA decomposition}

\item{U.hat}{Right singular matrix of L.burnin}

\item{V.hat}{left singular matrix of L.burnin}

\item{sigma.hat}{singular values of L.burnin}

\item{lambda1, }{lambda2 tuning parameters}

\item{decomp}{if TRUE, carry batch rPCA decomposition on N matrix}
}
\value{
S and L vectors for sample in question
}
\description{
Function that calls the online rPCA on sample under question and updates subspace
basis and does decomposition
}
\author{
Aditya Deshpande
}
